Now using the zQ class
Extend NodeList to add Array functions, making it not matter
if working with and array or node list.

zQ has class functions (add/remove), constructor, listeners...

Moved all zQ class to end, and should be ready to be moved to some external file if
I should choose to do so.