Return sanity, fewer exits, and do not store database crediential (read direct from connect.ini)
Removed all private vars relating to login. Now use a two vars within
the scope of __construct. One is the contents of connect.ini, the other is
a string to pass when constructing new PDO.

Fewer exits/dies (only exit on fail to connect), and use try/catch to avoid
errors containing database credentials.

Removed all DB_ prefixes for methods.

slq_table now tries to avoid a second query and gets table headers from
results of first query, if any. If it is an empty array, then it does the
second query to get the headers.