Handle ajax requests for home, posts, and tags. Removed $_POST[href].
Check if anything is passed in $_REQUEST and if this is a redirect. If neither of
these are true, it is an ajax request for the home page, so send that.

Next, check if this is a redirect with status 200. If so, convert the REDIRECT_URL
to an array by expoding it with '/'. Determine what is requested with 
switch($path[0]).

If none of these are true, continue as we did before.