urlencode(trim($tag)) to make sure that a tag of ' regular expression' works.
That way, we can get it back to the original value more easily using urldecode().

Otherwise, we were trying to get a tag of 'regular expression' from 
'-regular-expression'.

Since this is handled via urldecode when exploding $_SERVER[REDIRECT_URL],
there is no longer any need to store the converted tag as a seperate variable.