New case for edit_post in form switch. And restore database in action switch
Requires login as admin and checks nonce.
If these tests pass, uses a prepared statement to update the post, similar
to the same way posts are created to begin with.

Restoring the database requires login as admin.
If admin, check if {$connect->site}.sql exists and read its content
(file_get_contents does both of these) and execute the return from this in
$DB->query.

Notify on success of failure.