Updated restore() and added dump() method.
If restore() is not given the filename, get the database name from connect.ini
and use that as the filename instead.

file_get_contents() returns false if file cannot be read, so add if/else
conditions on its return. If $sql (will be true if it is the contents), continue
to executing query($sql).

PDO::query() returns false if the query fails, return $this->pdo->query($sql)
if we make it this far.

Returns false if file cannot be read or if query fails. Otherwise, returns
whatever the results of the query are, which evaluates as true.

dump() method gets connection information from connect.ini and exec()
to exec("mysqldump ....").
The command is built based on information in connect.ini, which means
that it adds -h $connect->server to the command if not using localhost.
Will want to add ability to change ports as well.