Requrie login now has more useful handling of role checking.
If not logged in, switch on $exit as before, but do the same for 403, '403',
and exit.

If logged in and $role is set, use array_search to compare $role with $login->role.
If either are not in the array, notify that the role checking was invalid.

If $required_level is greater than $user_level, notify that the user does not
have permission to perform the operation.

Returns true if logged in and $role is either not set or not greater than the
user's role.

Add attribute value=nonce to [name=nonce].

This should update any hidden inputs for nonce as well as the sessionStorage