Updated wywiwyg editor for use with document.execCommand()
No longer use getSelection() and substrings. Instead, use the built-in API
for contenteditable.

data-editor-command will be passed as the first argument for execCommand,
null will always be passed as the second, and the third argument can be one
of several things. If data-prompt is set, it will be the return from a prompt
created, using data-prompt as its message (for links and such). If
data-editor-value is set, it will use that instead. Finally, if none of these are
true, it will be null.

Added several more menuitems and menus, attempting to make better use
of the full set of features.

See https://developer.mozilla.org/en-US/docs/Midas