Now with RSS Feeds too!
SImilar to sitemaps, there is now a function called update_rss in custom.php
that does a few database queries to generate requierd information, and
then uses a template to create items. All of this, is saved to a file using fopen,
fputs, and finally fclose.

Added necessary <link> to head, and ignore *.rss in .gitignore