Merge <form>.ajaxSubmit() into ajax().
Removed <form>.ajaxSubmit() and updated ajax() to accomidate the change.

ajax() can be given data.form as either a string (name of the form) or
an HTML Form Element.

It will then use FormData, appending the form's name, and anything with
data-input, as ajaxSubmit() would.

Only set Content-Type header if not using FormData(), as FormData sets
it's own boundaries with multi-part form data type, and we have no access
to the boundary it sets, which causes the server to be unable to interpret the
data.

Updated bootstrap() in custom.js to reflect this, and also added support
for data-dropzone on inputs, which sets the value of the input to the
content of a dropzone set on the element identified by input's data-dropzone
(does document.querySelector())