Add `created` (datetime created) when creating new posts.
Also, add new post links to both <nav> & <menu label="Posts">

And, added "keep-alive" to actions. Intended to keep user logged in while
writing longer posts. Prior to this, long periods of staleness of session on
the server would allow the session to expire, causing the user to be logged
out unknowingly, which caused submission of new post to fail.