Keep-alive requests and drag n' drop on creating/editing posts.
Use an ajax request with requst: action=keep-alive to avoid session timing
out and user being logged out when creating or editing posts.

Drag n' Drop support. Drag in a file (image) while editing or creating a post
and it was be appended to the content. Still a work in progress.

Create events for load, dragover, dragend, drop. Add progress bar while
file is being read and remove upon completion.