New method and function to handle scrolling for handleJSON
If $resp->scrollTo($sel[, $nth]) is given to handleJSON, it will do
document.querySelectorAll($sel)[$nth].scrollIntoView(). Need to use $nth (
default is 0) in cases where :nth-of-type($nth) would not work due to it not
handling larger numbers. In cases where you would use ('li:nth-of-type(600)'),
it does not seem to work, so use ('li', 600) instead.