Nearly complete on install form handler
Creates user and sets up database. Multiple notifications along
the way if something goes wrong.

Need to add site specific information to head, such as author, title,
etc.

Also, need to create a default user.

Lastly, the default database needs to be created. Tested and working
with
a mysqldump of my actual database, but one needs to be created with some
Lorem Ipsum and such.